# Add local caching of `ColorSpace`s, by name, in `PartialEvaluator.getOperatorList` (issue 2504)